bit of clean up
